Decentralize importStataus
